requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.7.2
seaborn==0.13.2
PyYAML==6.0.2
missingno==0.5.2
scipy==1.14.1